9/10 update layout create quote

+++ lib/model/model_input_quote_detail.dart
@@ -7,7 +7,7 @@ class InputQuoteDetail {
   String? container;
   String? inGateDate;
   String? damageDetail;
-  int? quantity;
+  double? quantity;
   String? dimension;
   double? length;
   double? width;
