+++ lib/screen/quote/add_edit_quote/widget/port_quote/widget/import_button.dart
@@ -24,7 +24,7 @@ class _ImportButtonState extends State<ImportButton> {
     String? container;
     String? inGateDate;
     String? damageDetail;
-    int? quantity;
+    double? quantity;
     String? dimension;
     double? length;
     double? width;
